requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API clients
kaggle>=1.5.0
requests>=2.31.0

# LLM and MCP
openai>=1.0.0
langchain>=0.1.0
langgraph>=0.0.20
autogen>=0.2.0
transformers>=4.35.0
torch>=2.0.0

# Evaluation and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0

# Statistical validation
scipy>=1.11.0
scikit-learn>=1.3.0

